Moved LibraryTab to its own package
