+++ src/main/java/org/jabref/gui/dialogs/AutosaveUiManager.java
@@ -1,8 +1,8 @@
 package org.jabref.gui.dialogs;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.LibraryTab;
 import org.jabref.gui.exporter.SaveDatabaseAction;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.model.database.event.AutosaveEvent;
 import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.preferences.PreferencesService;
