+++ src/main/java/org/jabref/gui/exporter/SaveAllAction.java
@@ -3,8 +3,8 @@ package org.jabref.gui.exporter;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.Globals;
 import org.jabref.gui.JabRefFrame;
-import org.jabref.gui.LibraryTab;
 import org.jabref.gui.actions.SimpleCommand;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.preferences.PreferencesService;
