+++ src/main/java/org/jabref/gui/importer/GenerateEntryFromIdAction.java
@@ -3,10 +3,10 @@ package org.jabref.gui.importer;
 import java.util.Optional;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.externalfiles.ImportHandler;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.gui.util.BackgroundTask;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.importer.CompositeIdFetcher;
