+++ src/main/java/org/jabref/gui/importer/GenerateEntryFromIdDialog.java
@@ -6,9 +6,9 @@ import javafx.scene.control.DialogPane;
 import javafx.scene.control.TextField;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.icon.IconTheme;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.util.FileUpdateMonitor;
