+++ src/main/java/org/jabref/gui/importer/actions/GUIPostOpenAction.java
@@ -1,6 +1,6 @@
 package org.jabref.gui.importer.actions;
 
-import org.jabref.gui.LibraryTab;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.logic.importer.ParserResult;
 
 /**
