+++ src/main/java/org/jabref/gui/importer/actions/MergeReviewIntoCommentAction.java
@@ -2,7 +2,7 @@ package org.jabref.gui.importer.actions;
 
 import java.util.List;
 
-import org.jabref.gui.LibraryTab;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.migrations.MergeReviewIntoCommentMigration;
 import org.jabref.model.entry.BibEntry;
