+++ src/main/java/org/jabref/gui/search/RebuildFulltextSearchIndexAction.java
@@ -3,9 +3,9 @@ package org.jabref.gui.search;
 import java.io.IOException;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.SimpleCommand;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.gui.util.BackgroundTask;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.l10n.Localization;
