+++ src/main/java/org/jabref/gui/undo/UndoRedoAction.java
@@ -5,11 +5,11 @@ import javax.swing.undo.CannotUndoException;
 
 import org.jabref.gui.DialogService;
 import org.jabref.gui.JabRefFrame;
-import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.actions.StandardActions;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.logic.l10n.Localization;
 
 import org.slf4j.Logger;
