+++ src/test/java/org/jabref/gui/importer/NewEntryActionTest.java
@@ -3,8 +3,8 @@ package org.jabref.gui.importer;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.EntryTypeView;
 import org.jabref.gui.JabRefFrame;
-import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
+import org.jabref.gui.library.LibraryTab;
 import org.jabref.gui.util.OptionalObjectProperty;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.types.EntryType;
